move gameobject and gamestate
