+++ src/server/CMakeLists.txt
@@ -20,4 +20,5 @@ target_link_libraries(${TARGET_NAME} PRIVATE
     glfw
     libglew_static
     asio
+    assimp
 )
