+++ src/shared/CMakeLists.txt
@@ -13,4 +13,5 @@ target_link_libraries(${LIB_NAME} PRIVATE
     glfw
     libglew_static
     asio
+    assimp
 )
